Add ext/marshmallow/common.py
common.py contains resolve_schema_instance and resolve_schema_cls.

This commit removes a duplication (get_schema_class / resolve_schema_cls)
and solves a circular import issue.
